feat: add emoticone button. fix:panic error, scroll jumping, frequent re-rendering and code improvement
