+++ share/dashboard_react/package.json
@@ -29,7 +29,9 @@
     "recharts": "^2.12.7",
     "redux-logger": "^3.0.6",
     "remark-gfm": "^4.0.0",
-    "sass": "^1.77.8"
+    "sass": "^1.77.8",
+    "@emoji-mart/react": "^1.1.1",
+    "react-emoji-render": "1.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
