+++ share/dashboard_react/src/Pages/Mattermost/styles.module.scss
@@ -62,7 +62,7 @@
             padding: 0.75rem;  // 12px / 16 = 0.75rem
             font-size: 0.875rem;  // 14px / 16 = 0.875rem
             margin-top: 0.5rem;  // 8px / 16 = 0.5rem
-            min-height: 7.5rem;  // 120px / 16 = 7.5rem
+            min-height: 3.75rem;  // 120px / 16 = 7.5rem
             resize: vertical;
             border: none;
